feat: Use int64_t for chunk shapes and validate dimensions
This commit refactors the file format and I/O layers to use `int64_t` instead of `uint32_t` for chunk shape dimensions. This change future-proofs the format, allowing for dimension sizes that exceed the `uint32_t` limit.

Key changes:
- The `Chunk` struct in `cdd_file_format.h` now uses `memory::vector<int64_t>` for `shape_`.
- Serialization and deserialization logic in `cdd_file_format.cpp` and `data_writer.cpp` has been updated to handle the 8-byte integer type.
- All I/O and compression APIs (`DataWriter`, `DataReader`, `DataCompressor`, `DataExtractor`) have been updated to accept and process `int64_t` shapes.
- Added robust validation to reject negative shape dimensions at the API boundary in `DataWriter` and `DataCompressor`.
- Added safety checks in `DataReader` and `DataExtractor` to handle potentially corrupt files with negative dimensions gracefully.
- Updated the test suite in `cdd_file_tests.cpp` to use `int64_t` for shape definitions and comparisons.
- Added a new test case, `WriterRejectsNegativeShape`, to explicitly verify that the writer rejects invalid negative dimensions.

+++ src/data_io/data_writer.h
@@ -90,7 +90,7 @@ public:
      * @return The index of the newly appended chunk on success, or an error string.
      */
     std::expected<size_t, std::string> append_chunk(ChunkDataType type, DType dtype, uint64_t flags,
-                                                  std::span<const uint32_t> shape, std::span<const std::byte> data);
+                                                  std::span<const int64_t> shape, std::span<const std::byte> data);
 
     /**
      * @brief Flushes any buffered data to the underlying storage.
